feat(sdk): move the descriptor language into @vyuhlabs/dxkit-sdk
HttpFlowSupport, FileRouteSupport, and ModelSchemaSupport relocate
verbatim to packages/dxkit-sdk/src/descriptors.ts — they are the
declarative language packs AND rung-4 plugin dialects speak, so their
one definition belongs on the frozen surface. src/languages/types.ts
imports and re-exports them, so no other consumer changes an import.
